requirements.txt
numpy
pandas
fasttext
nltk
xlrd
tweet-preprocessor